requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
bsedata==0.5.0
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
mysqlclient==2.1.0
nsepython==0.0.972
numpy==1.21.4
pandas==1.3.4
pkg_resources==0.0.0
pycodestyle==2.7.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.0.0
Unipath==1.1
urllib3==1.26.7
whitenoise==5.3.0
